refactor: move command descriptions out of TUI module
