+++ src/cli/commands/package/command.tsx
@@ -1,5 +1,5 @@
+import { COMMAND_DESCRIPTIONS } from '../../constants';
 import { getErrorMessage } from '../../errors';
-import { COMMAND_DESCRIPTIONS } from '../../tui/copy';
 import { handlePackage, loadPackageConfig } from './action';
 import type { Command } from '@commander-js/extra-typings';
 import { Text, render } from 'ink';
